Добавлена обработка исключений + UserService + FilmService,

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -13,8 +13,9 @@ import java.util.Set;
 @AllArgsConstructor
 @Builder
 public class Film {
+
     @PositiveOrZero
-    private int id;
+    private Long id;
     @NotBlank
     private String name;
     @NotNull
@@ -36,4 +37,27 @@ public class Film {
     public int getLikesQuantity() {
         return likes.size();
     }
+//    @PositiveOrZero
+//    private long id;
+//    @NotBlank
+//    private String name;
+//    @NotNull
+//    @Size(min = 1, max = 200)
+//    private String description;
+//    private LocalDate releaseDate;
+//    @Positive
+//    private long duration;
+//    private Set<Long> likes;
+//
+//    public void addLike(Long userId) {
+//        likes.add(userId);
+//    }
+//
+//    public void removeLike(Long userId) {
+//        likes.remove(userId);
+//    }
+//
+//    public int getLikesQuantity() {
+//        return likes.size();
+//    }
 }
